add new branch - dev-planetas

+++ src/main/java/com/mercado/libre/model/Punto.java
@@ -1,33 +1,35 @@
 package com.mercado.libre.model;
 
+import java.math.BigDecimal;
+
 /**
  *
  */
 public class Punto {
 
-    private Double x;
-    private Double y;
+    private BigDecimal x;
+    private BigDecimal y;
 
-    public Punto(Double x , Double y ){
+    public Punto(BigDecimal x , BigDecimal y ){
         this.x = x;
         this.y = y;
     }
 
     public Punto(){}
 
-    public Double getX() {
+    public BigDecimal getX() {
         return x;
     }
 
-    public void setX(Double x) {
+    public void setX(BigDecimal x) {
         this.x = x;
     }
 
-    public Double getY() {
+    public BigDecimal getY() {
         return y;
     }
 
-    public void setY(Double y) {
+    public void setY(BigDecimal y) {
         this.y = y;
     }
 }
